docker-compose/requirements.txt
django==4.2.11
flake8==6.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
django-split-settings==1.3.2
django-extensions==3.2.3
uWSGI==2.0.20 
ipython==8.18.1